+++ src/renderer/src/components/player/src/core/index.ts
@@ -1,6 +1,7 @@
 import ArtPlayerAdapter from './artplayer';
-import DPlayerAdapter from './dplayer';
-import NPlayerAdapter from './nplayer';
+// import DPlayerAdapter from './dplayer';
+// import NPlayerAdapter from './nplayer';
 import XgPlayerAdapter from './xgplayer';
 
-export { ArtPlayerAdapter, DPlayerAdapter, NPlayerAdapter, XgPlayerAdapter };
+// export { ArtPlayerAdapter, DPlayerAdapter, NPlayerAdapter, XgPlayerAdapter };
+export { ArtPlayerAdapter, XgPlayerAdapter };